added a check for deleting only single eligibilities when there are no childs.  and no deleting of childs.
git-tfs-id: [https://tfs.caes.ucdavis.edu:8088/tfs/DefaultCollection]$/EL/Trunk/EL;C7661

+++ EligibilityList/Controllers/EligibilityController.cs
@@ -310,6 +310,12 @@ namespace EligibilityList.Controllers
                 return RedirectToAction("Index");
             }
 
+            if (el.OriginalEligibility != null || Repository.OfType<Eligibility>().Queryable.Where(a=>a.OriginalEligibility == el).Any())
+            {
+                Message = "Eligibility cannot be deleted, please complete any pending action first.";
+                return this.RedirectToAction(a => a.Edit(id));
+            }
+
             _eligibilityRepository.Remove(el);
 
             return RedirectToAction("Index");
